LazyFS: Config option 'sync_after_rename' should be set to false by default
This option was added to the configuration file as part of solving a missing fsync from PostgreSQL (statistics file). However, if DBs are fsyncing the most important filescorrectly, then this option should be set to false. By having it set as 'true', LazyFS was performing implicit fsyncs and DB tests were not realistic.